Implemented Camera2 API, including its Activity Read more...
To do that, I needed to make a Thread class which takes the
AndroidCameraApi Activity context and takes photos each x seconds, we need this
thread running in background taking photos without the user knowing as
we want to process those photos "silently".

There are some thread exceptions when AndroidCameraApi Activity is
closed, I tried to figure out what is going on, but finally I handled it
with a try catch block since it appears only once. As I saw it is
because I can't find a good method to shutdown the thread when the user
presses the back button. Maybe later we can dig more and try to figure
out how to do it properly.